requirements.txt
pytest>=8.0.1
jinja2>=3.1.3
python-dotenv>=1.0.1